Create CoulombMatrix-Generator
CoulombMatrix-Generator creates a list of CMs that are invariant to SMILES input. For example, the code has 6x SMILES codes for ethanol, and they have identical CMs. This is because SMILES codes are canonicalized before the matrix is created. Importantly, only heavy atoms are considered by this program; i.e., any explicit hydrogen in the SMILES code is removed.